Abandoning websockets for CRUD operations with posts and switching to viewsets.
It was decided to switch from using websocket for CRUD operations with posts to using veiwset. Websocket will be used as a signal for frontend to add new posts (button "show new n-posts").

Based on PostConsumers, ViewSets for CRUD operations besides list have been developed. For list, pagination needs to be checked. The methods need to be optimised, especially retrieve and list