Like/dislike system and PostDetail
On the backend: the display of the title and decription fields in the admin panel have been changed using the forms widget (the fields display multiple rows as TextField)

On the frontend: the mechanism of requests for likes/dislikes has been tested and the PostDetail page has been added.

Need to do: rewrite the view on the backend to display reactions to the post.